Struct matrix_display::cell::Cell

pub struct Cell<T> where T: Clone, T: ToString {
    pub value: T,
    pub color: AnsiColor,
}

A Matrix Cell that owns some data, a background color and a foreground color

The colors are stored by their ansi codes in an AnsiColor struct The data can be any type that is clonable and converts to a string

Example (a chess game's representation):

let board = vec!['♜', '♞', '♝', '♛', '♚', '♝', '♞', '♜',
                 '♟', '♟', '♟', '♟', '♟', '♟', '♟', '♟',
                 ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ',
                 ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ',
                 ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ',
                 ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ',
                 '♖', '♘', '♗', '♕', '♔', '♗', '♘', '♖',
                 '♙', '♙', '♙', '♙', '♙', '♙', '♙', '♙']
    .iter()
    .enumerate()
    .map(|(i, x)| {
       let ansi_fg = 33; // pawns in blue
        let mut ansi_bg = 0;
        if i % 2 + (i / 8) % 2 == 1 {
            ansi_bg = 7;
        }
        matrix_display::cell::Cell::new(x.clone(), ansi_fg, ansi_bg)
        })
    .collect::<Vec<_>>();

Fields
value: T color: AnsiColor
Methods

impl<T> Cell<T> where T: Clone, T: ToString

fn new(val: T, fg: u8, bg: u8) -> Cell<T>

Construct a cell with any data, and two ansi codes for the foreground and the background colors