Module matrix_display::matrix

Provides Matrix which stores a matrix of arbitrary data

Data is stored as a vector in row major order and a usize representing the number of columns Data can be accessed for reading through the enumerate_cells function That function returns a vector of a clone of each cell and its Position

Modules

position
A simple human readable way of defining a position within a box

Structs

Matrix
Stores a matrix of arbitrary data