Enum matrix_display::matrix::position::Position

pub enum Position {
    Top,
    Left,
    Right,
    Bottom,
    TopLeft,
    TopRight,
    BottomLeft,
    BottomRight,
    Middle,
}

Enum that describes a position within a rectangle

The position can be accessed at two levels of granularity: The values of the enum denote the exact position (i.e: BottomLeft) Helper functions help determine wether that position is part of a given border (i.e: BottomLeft is both part of bottom() and left())

Variants
TopLeftRightBottomTopLeftTopRightBottomLeftBottomRightMiddle
Methods

impl Position

fn top(&self) -> bool

Is it part of the top row?

fn left(&self) -> bool

Is it part of the leftmost column?

fn right(&self) -> bool

Is it part of the rightmost column?

fn bottom(&self) -> bool

Is it part of the bottom row?